pyproject.toml
[project]
name = "langgraph-study1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.1.0",
    "grandalf>=0.8",
    "langchain>=0.3.24",
    "langchain-community>=0.3.23",
    "langchain-google-genai>=2.1.3",
    "langchain-groq>=0.3.2",
    "langchain-ollama>=0.3.2",
    "langchainhub>=0.1.21",
    "langgraph>=0.4.0",
    "langsmith>=0.3.38",
    "python-dotenv>=1.1.0",
]
